armies are locked down to their owners

+++ campaignManager/campaignManager/armies/views.py
@@ -1,18 +1,23 @@
 from django.shortcuts import render
 from campaignManager.armies.models import *
-from django.shortcuts import render, render_to_response, get_object_or_404, redirect
+from django.shortcuts import render, render, get_object_or_404, redirect
 from django.core.context_processors import csrf
 from django.template.context import RequestContext
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.contrib.messages import constants as message_constants
 
+MESSAGE_TAGS = {message_constants.ERROR: 'danger'}
 
 # Create your views here.
 def detail(request, pk):
     army = get_object_or_404(Army, pk=pk, public_list=True)
-    return render_to_response('detail.html', {
+
+    return render(request, 'detail.html', {
         'request': request,
         'user': request.user,
-        'army': army
+        'army': army,
+        'editable': army.is_owned_by(request.user)
     })
 
 def index(request, slug=None):
@@ -21,7 +26,7 @@ def index(request, slug=None):
     else:
         armies = Army.objects.filter(public_list=True)
     
-    return render_to_response('index.html', {
+    return render(request, 'index.html', {
         'request': request,
         'user': request.user,
         'armies': armies
@@ -29,13 +34,19 @@ def index(request, slug=None):
     
 @login_required
 def edit(request, pk=None):
-    context = RequestContext(request, {
-        'request': request,
-        'user': request.user
-    })
+    delete = None # Initial state
     
-    if pk: army = get_object_or_404(Army, pk=pk);
-    
+    if pk: 
+        army = get_object_or_404(Army, pk=pk);
+        if army.user and not army.is_owned_by(request.user):
+            messages.add_message(
+                request, 
+                messages.ERROR, 
+                'Armies can only be edited by their owners'
+            )
+            return redirect('armies:detail', pk=army.pk)
+        delete = {'path': 'armies:delete', 'value': army.pk, }
+            
     if request.method == 'POST':
         if pk:
             form = ArmyForm(request.POST, instance=army)
@@ -45,21 +56,29 @@ def edit(request, pk=None):
             army = form.save(commit=False)
             army.user = request.user
             army.save()
-            return redirect('armies:detail', pk=army.id)
+            messages.add_message(request, messages.SUCCESS, 'Save Successful' )
+            
+        return redirect('armies:detail', pk=army.pk)
         
     elif pk:
         form = ArmyForm(instance=army)
     else:
         form = ArmyForm()        
             
-    return render_to_response('form.html', {
+    return render(request, 'form.html', {
         'form': form,
         'request': request,
         'user': request.user,
-    }, context_instance=context)
+        'delete': delete,
+    })
     
 @login_required
 def delete(request, pk):
     army = get_object_or_404(Army, pk=pk)
-    army.delete()
-    return redirect('armies:index')
+    if army.is_owned_by(request.user):
+        army.delete()
+        messages.add_message(request, messages.SUCCESS, 'Delete successful')
+        return redirect('armies:index')
+    else:
+        messages.add_message(request, messages.ERROR, 'Armies can only be deleted by their owners')
+        return redirect('armies:detail', army.pk);
